added running

+++ server/src/game_state.rs
@@ -6,8 +6,8 @@ pub struct PlayerData {
     pub is_running: bool,
     pub next_move_direction: Option<u8>,
     pub direction: u16,
-    pub x: i32,
-    pub y: i32
+    pub x: f64,
+    pub y: f64
 }
 
 #[derive(Clone,Debug)]
